### 2020-06-11 ccbebf4
bibtex normalization
