+++ paper/paper.bib
@@ -1,62 +1,62 @@
 
 @inproceedings{khan_similarity_2016,
-	address = {Dhaka, Bangladesh},
-	title = {Similarity analysis of patients' data: {Bangladesh} perspective},
-	isbn = {9781509054213},
-	shorttitle = {Similarity analysis of patients' data},
-	url = {http://ieeexplore.ieee.org/document/7835390/},
-	doi = {10.1109/MEDITEC.2016.7835390},
-	urldate = {2020-06-11},
-	booktitle = {2016 {International} {Conference} on {Medical} {Engineering}, {Health} {Informatics} and {Technology} ({MediTec})},
-	publisher = {IEEE},
-	author = {Khan, Shahidul Islam and Hoque, Abu Sayed Md. Latiful},
-	month = dec,
-	year = {2016},
-	pages = {1--5}
+  address = {Dhaka, Bangladesh},
+  title = {Similarity analysis of patients' data: {Bangladesh} perspective},
+  isbn = {9781509054213},
+  shorttitle = {Similarity analysis of patients' data},
+  url = {http://ieeexplore.ieee.org/document/7835390/},
+  doi = {10.1109/MEDITEC.2016.7835390},
+  urldate = {2020-06-11},
+  booktitle = {2016 {International} {Conference} on {Medical} {Engineering}, {Health} {Informatics} and {Technology} ({MediTec})},
+  publisher = {IEEE},
+  author = {Khan, Shahidul Islam and Hoque, Abu Sayed Md. Latiful},
+  month = dec,
+  year = {2016},
+  pages = {1--5}
 }
 
 @article{pollock_automatic_1984,
-	title = {Automatic spelling correction in scientific and scholarly text},
-	volume = {27},
-	issn = {00010782},
-	url = {http://portal.acm.org/citation.cfm?doid=358027.358048},
-	doi = {10.1145/358027.358048},
-	number = {4},
-	urldate = {2020-06-11},
-	journal = {Communications of the ACM},
-	author = {Pollock, Joseph J. and Zamora, Antonio},
-	month = apr,
-	year = {1984},
-	pages = {358--368}
+  title = {Automatic spelling correction in scientific and scholarly text},
+  volume = {27},
+  issn = {00010782},
+  url = {http://portal.acm.org/citation.cfm?doid=358027.358048},
+  doi = {10.1145/358027.358048},
+  number = {4},
+  urldate = {2020-06-11},
+  journal = {Communications of the ACM},
+  author = {Pollock, Joseph J. and Zamora, Antonio},
+  month = apr,
+  year = {1984},
+  pages = {358--368}
 }
 
 @inproceedings{broder_resemblance_1998,
-	address = {Salerno, Italy},
-	title = {On the resemblance and containment of documents},
-	isbn = {9780818681325},
-	url = {http://ieeexplore.ieee.org/document/666900/},
-	doi = {10.1109/SEQUEN.1997.666900},
-	urldate = {2020-06-11},
-	booktitle = {Proceedings. {Compression} and {Complexity} of {SEQUENCES} 1997 ({Cat}. {No}.{97TB100171})},
-	publisher = {IEEE Comput. Soc},
-	author = {Broder, A.Z.},
-	year = {1998},
-	pages = {21--29}
+  address = {Salerno, Italy},
+  title = {On the resemblance and containment of documents},
+  isbn = {9780818681325},
+  url = {http://ieeexplore.ieee.org/document/666900/},
+  doi = {10.1109/SEQUEN.1997.666900},
+  urldate = {2020-06-11},
+  booktitle = {Proceedings. {Compression} and {Complexity} of {SEQUENCES} 1997 ({Cat}. {No}.{97TB100171})},
+  publisher = {IEEE Comput. Soc},
[…]
