### 2020-06-18 #2327 @Kevin-Mattheus-Moerman
@Yomguithereal we are currently trying to find an editor to handle this submission. In the meantime please check our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) and amend your paper where needed. In particular you should cite "A list of key references, including to other software addressing related needs".
With sentences in your paper like this:
`Thus, even if Talisman obviously provides state-of-the-art functions that are still being used in an industrial context, it also aims at being a safe harbor for less known or clunkier techniques, for historical and archival purposes.`
You have the opportunity to cite these works/methods you are referring to. Please also briefly discuss other existing software projects addressing similar functionality.
After adding citations please do check that you've added DOI's where possible.
You can call `@whedon generate pdf` and `@whedon check references` here yourself too.
