### 2020-06-22 9bd0055
chore(types): for inflectors / keyers
