### 2020-06-22 b050cb1
fix(types): keyers & tokenizers’s fingerprint
