+++ index.d.ts
@@ -6,28 +6,11 @@ declare module "talisman/inflectors/spanish" {
 }
 
 declare module "talisman/keyers/fingerprint" {
-  // represented type of options parameter
-  export interface TokenizerOptions {
-    digits: boolean;
-    minTokenSize: number;
-    ngrams: boolean;
-    sort: boolean;
-    split: string[] | null;
-    stopwords: string[] | null;
-  }
-
-  export function createTokenizer(options?: TokenizerOptions): NgramFunction;
-  export default NgramFunction;
+  export default function fingerprint(str: string): string;
 }
 
 declare module "talisman/keyers/ngram-fingerprint" {
-  const ngrams: NgramFunction;
-  export default ngrams;
-
-  // popular aliases
-  export const bigrams: NgramFunction;
-  export const trigrams: NgramFunction;
-  export const quadrigrams: NgramFunction;
+  export default function ngramFingerprint(n: number, sequence: string): string;
 }
 
 declare module "talisman/keyers/name-sig" {
@@ -42,3 +25,27 @@ declare module "talisman/keyers/skeleton" {
   export default function skeleton(str: string): string;
 }
 
+declare module "talisman/tokenizers/fingerprint" {
+  // represented type of options parameter
+  export interface TokenizerOptions {
+    digits: boolean;
+    minTokenSize: number;
+    ngrams: boolean;
+    sort: boolean;
+    split: string[] | null;
+    stopwords: string[] | null;
+  }
+
+  export function createTokenizer(options?: TokenizerOptions): NgramFunction;
+  export default NgramFunction;
+}
+
+declare module "talisman/tokenizers/ngram-fingerprint" {
+  const ngrams: NgramFunction;
+  export default ngrams;
+
+  // popular aliases
+  export const bigrams: NgramFunction;
+  export const trigrams: NgramFunction;
+  export const quadrigrams: NgramFunction;
+}
