### 2020-06-22 #2327 @Yomguithereal
> @Yomguithereal we are currently trying to find an editor to handle this submission. In the meantime please check our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) and amend your paper where needed. In particular you should cite "A list of key references, including to other software addressing related needs".
> With sentences in your paper like this:
> `Thus, even if Talisman obviously provides state-of-the-art functions that are still being used in an industrial context, it also aims at being a safe harbor for less known or clunkier techniques, for historical and archival purposes.`
> You have the opportunity to cite these works/methods you are referring to. Please also briefly discuss other existing software projects addressing similar functionality.
> After adding citations please do check that you've added DOI's where possible.
> You can call `@whedon generate pdf` and `@whedon check references` here yourself too.

Hello everyone, sorry for the delay.

I have an extensive bibliography [here](https://github.com/Yomguithereal/talisman/blob/master/paper/paper.bib) but I did not find how to automatically generate it under the `References` section of the abstract. Should I only pick a limited amount?

Also, regarding links to similar software endeavors can I just edit my repo and then ping you when this is done?

Finally, regarding DOIs I tried as much as possible to get them but I am afraid that most of my references don't have one. Or maybe there is a way to find DOI easily (most of my bibtex comes straight from google scholar & zbib).

Have a good day and thanks for agreeing to review the library.
