### 2020-06-23 #2327 @arfon
> I have an extensive bibliography here but I did not find how to automatically generate it under the References section of the abstract. Should I only pick a limited amount?

@Yomguithereal - You need to cite them directly please for them to be rendered into the document (You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)).
