### 2020-06-23 #2327 @danielskatz
And just to add, JOSS does not publish papers that are surveys, so the bibliography (and ideally the .bib file due to a limitation in our tooling) should only contain items that are referenced in the paper.
