### 2020-06-23 #2327 @Yomguithereal
All the references existing in the .bib files are papers describing methods actually implemented by the library. So do you agree I should list every reference in the abstract under `References` then? If so I will probably create a script to do so if such script does not already exist.
