### 2020-06-23 #2327 @danielskatz
In my opinion, if you describe the method in the paper, it would be reasonable to cite a reference for that method.  The best thing to do is to look at your paper, and decide where you need to place citations to references, then to include those references in the .bib file.  Pandoc as called by whedon will then build the reference list in the pdf.
