### 2020-06-23 #2327 @Yomguithereal
The paper does not describe a specific method but rather explains the scope and stake of the library as an aggregation of implementations of various fuzzy matching etc. methods. The lib implements a large amount of different methods/functions etc. which are nearly exhaustively referenced by papers, collected in the current bibliography. So I guess I am misunderstanding the framing the paper should have. For instance, if we take the example of scikit-learn, do you expect to have few references linking towards machine learning as a field or rather the exhaustive list of papers that have been used to implement their large collection of algorithms?
