### 2020-06-23 #2327 @danielskatz
I think my answer is yes.  An academic paper on scikit-learn might go into a lot of detail about all the methods, with citations to references for each method, but this is probably not appropriate for a JOSS paper.

See https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain for guidance on what a JOSS paper should contain.
