### 2020-06-26 #2327 @Fil
I'm ready to start reviewing *talisman* but I can't find the usual checklist? has the review process changed in some mysterious way since I last did [a review](https://github.com/openjournals/joss-reviews/issues/1628)? What should I do?
