### 2020-06-29 #2405 @whedon
[REVIEW]: Talisman: a JavaScript archive of fuzzy matching, information retrieval and record linkage building blocks
**Submitting author:** @Yomguithereal (<a href="http://orcid.org/0000-0003-4916-8472">Guillaume Plique</a>)
**Repository:** <a href="https://github.com/Yomguithereal/talisman" target ="_blank">https://github.com/Yomguithereal/talisman</a>
**Version:** 1.1.3-zenodo
