### 2020-06-30 Yomguithereal/talisman#172 @Fil
JOSS review
All set [for my review](https://github.com/openjournals/joss-reviews/issues/2405) except this point:
* the guidelines ask us to verify _**State of the field:** Do the authors describe how this software compares to other commonly-used packages?_ => This seems to be lacking at the moment. (Not sure if that is what you mean in #170)

(It would also be great to have the possibility of a root entry point #10, but not a blocker: the curent installation process is functional and well documented).
