### 2020-06-30 Yomguithereal/talisman#172 @Yomguithereal
Hello @Fil. #170 is not related to this I think. I was just hinting at the fact that this library can be used to perform actual precision/recall benchmarks etc. of various methods, something which is more rarely done in this field than in others because of the lack of unified interface such as what this lib offers. (This is off-topic but I am currently building a database of name variants to run such a benchmark with phonetic algorithms).

As someone indicated also in the pre-review, the paper lacks of pointers towards other similar libraries, in JS or other languages and so I intend to add some links soon. I just want to know if I can do this now or if I should wait for you and the other reviewer to first iterate once on the review.

For instance I would like to list [abydos](https://pypi.org/project/abydos/) or things like [java-string-similarity](https://github.com/tdebatty/java-string-similarity) and OpenRefine etc. Would this fit the bill? How many libs/tools should I link to?
