### 2020-06-30 Yomguithereal/talisman#172 @Fil
Feel free to develop your software when you want, there's no "review version" and in any case we can follow the changes. You should "describe how this software compares to other commonly-used packages", it's not for us to be normative about the way you do this (it can be anywhere between an off-hand comment and a full benchmark).
