### 2020-06-30 Yomguithereal/talisman#172 @Yomguithereal
What I don't know yet also is what to do with the exhaustive bibliography of papers implemented by the library. As I understood from the pre-review discussion, I should only list references mentioned by the paper, no? Is there anyway to give some value to the full bibliography?
