### 2020-06-30 Yomguithereal/talisman#172 @Fil
It would be great to have it as a page in the extensive documentation
https://yomguithereal.github.io/talisman/

-- Fil

On Tue, Jun 30, 2020 at 3:01 PM Guillaume Plique <notifications@github.com>
wrote:

> What I don't know yet also is what to do with the exhaustive bibliography
> of papers implemented by the library. As I understood from the pre-review
> discussion, I should only list references mentioned by the paper, no? Is
> there anyway to give some value to the full bibliography?
>
