### 2020-06-30 Yomguithereal/talisman#172 @Yomguithereal
> It would be great to have it as a page in the extensive documentation

That's a good idea. I will do that, thank you.
