### 2020-07-09 #2405 @kakiac
👋  @Fil many thanks for your comments - @yomguithereal can you let us know once you have addressed @Fil 's comments?

@atanikan let me know if you need any help with the reviewing process - the review criteria you need to follow are outlined here: https://joss.readthedocs.io/en/latest/review_criteria.html and a useful checklist is above (with comments and explanations here: https://joss.readthedocs.io/en/latest/review_checklist.html) feel free to check the boxes directly!

Many thanks all for all your hard work in improving the library and the paper,
