### 2020-07-09 #2405 @Yomguithereal
> @Yomguithereal can you let us know once you have addressed @Fil 's comments?

Sure, I will tell you when this is done. I first need to find a little bit of time to do so, hope there is no time limit on this :) Thanks for reviewing the library.
