### 2020-07-09 #2405 @atanikan
@kakiac Thanks for the reminder :)
@Yomguithereal The installation works. I checked some of your functions and they work well. Minor suggestions, Personally I'd find it more intuitive to do
`var talis = require('talisman');
talis.metrics.cosine([1,2],[3,4]);`
over
`var cos = require('talisman/metrics/cosine');
cos([1,2],[3,4]);`
I guess the [issue](https://github.com/Yomguithereal/talisman/issues/10) does speak to it.
A comparative benchmark study as mentioned [here](https://github.com/Yomguithereal/talisman/issues/170) which can also be added to your documentation website would be super useful.
