### 2020-07-09 #2405 @Yomguithereal
> I guess the issue does speak to it.

Yes the issue is indeed about this point. The reason why this was implemented thusly, in the beginning, is because it was written at a time where tree-shaking bundler were not yet a thing and since the library is really heavy on its own, it was the only way to make sure people would only require pieces of it on a need-to basis. But I agree this is somewhat unnecessary when using the library in node.js

I will add a centralized endpoint in a near future but I need some time to be sure I can do so while supporting ES6 imports, CommonJS & TypeScript which is unfortunately always a drag.

> A comparative benchmark study as mentioned here which can also be added to your documentation website would be super useful.

This issue (imprecisely) refer to the fact that you can use the library to benchmark several methods such as phonetic encoders against a database of, say, given names. I was not refering to performance benchmark against other similar libraries. Is that what you had in mind?
