### 2020-07-09 #2405 @atanikan
> I will add a centralized endpoint in a near future but I need some time to be sure I can do so while supporting ES6 imports, CommonJS & TypeScript which is unfortunately always a drag.

That's fair! I guess we can pass on it for now.

> This issue (imprecisely) refer to the fact that you can use the library to benchmark several methods such as phonetic encoders against a database of, say, given names. I was not referring to performance benchmark against other similar libraries. Is that what you had in mind?

Then yes, I meant performance benchmarking with other similar libraries. I understand Talisman aims to be an archive/collection of several text-based search/information retrieval techniques. And not exactly a brand new algorithm. I however feel some benchmarking, be it in terms of accuracy or speed of execution by taking a large dataset as an example, may prove to be useful for wider use and a greater acceptance of your library.
