### 2020-09-02 80c87bc
Fixed phonetics/french/fonem V-10 rule Related to #175
