+++ src/phonetics/french/fonem.js
@@ -38,7 +38,7 @@ const RULES = {
   'V-7': [/([^G])AY$/, '$1E'],
   'V-8': [/EUX$/, 'EU'],
   'V-9': [`EY(?=$|${C})`, 'E'],
-  'V-10': [`(${C})?Y(?!${V})`, '$1I'],
+  'V-10': [`(^|${C})Y(?=${V}|$)|(^|${V})Y(?=${C}|$)`, '$1$2I'],
   'V-11': [`(${V})I(?=${V})`, '$1Y'],
   'V-12': [`(${V})ILL`, '$1Y'],
   'V-13': [/OU(?=I(?!LL)|[AEOU])/g, 'W'],
