### 2020-09-02 9d78ab4
Simpler phonetics/french/fonem V-10 rule Related to #175
