### 2020-09-02 fec8bdb
Fixing phonetics/french/fonem C-28 rule Relate to #175
