+++ src/phonetics/french/fonem.js
@@ -80,7 +80,7 @@ const RULES = {
   'C-25': [/([AIOUY])W/g, '$1'],
   'C-26': [/X[CSZ]/g, 'X'],
   'C-27': [`(?:Z(?=${V})|(${C})Z(?=${C}))`, '$1S'],
-  'C-28': [`(?:([${CONSONANTS}CLS])\\1|(C)C(?!${V})|(.S)S(?!${V})|([^I]L)L)`, '$1$2$3$4'],
+  'C-28': [`(?:([${CONSONANTS}CLS])\\1|(C)C(?!${V})|(${C}S)S|(.S)S(?!${V})|([^I]L)L)`, '$1$2$3$4$5'],
   'C-28-bis': [/ILE$/, 'ILLE'],
   'C-29': [`(?:(ILS)|([CS]H)|([MN]P)|(R[CFKLNSX])|(${C})${C})$`, '$1$2$3$4$5'],
   'C-30': [/^(?:SINT?|SAINT?|SEIN|SEIM|CINQ?)/, 'ST-'],
