### 2020-10-19 #2405 @kakiac
👋 @Yomguithereal - hope you are well! I am just following up to see how things are progressing with implementing the reviewer's suggestions? Let me know if there is something I can do to help our end :) Hope you are well 👍
