### 2020-10-19 #2405 @Yomguithereal
Hello @kakiac. I should have some time thursday to act on those suggestions. So far I need to update the paper to change its framing a bit, add some references to similar tools and switch to a leaner bibliography (as opposed to the one containing all of the papers describing the implemented functions and methods). I also have some one related to the library's packaging. I would like to address those but later since it is very time consuming.
