### 2020-10-22 7bf24af
Adjusting paper   * Splitting algorithms.bib & paper.bib   * Citing some papers in the reference   * Adding a related works section
