+++ paper/paper.bib
@@ -1,84 +1,4 @@
 
-@inproceedings{khan_similarity_2016,
-  address = {Dhaka, Bangladesh},
-  title = {Similarity analysis of patients' data: {Bangladesh} perspective},
-  isbn = {9781509054213},
-  shorttitle = {Similarity analysis of patients' data},
-  url = {http://ieeexplore.ieee.org/document/7835390/},
-  doi = {10.1109/MEDITEC.2016.7835390},
-  urldate = {2020-06-11},
-  booktitle = {2016 {International} {Conference} on {Medical} {Engineering}, {Health} {Informatics} and {Technology} ({MediTec})},
-  publisher = {IEEE},
-  author = {Khan, Shahidul Islam and Hoque, Abu Sayed Md. Latiful},
-  month = dec,
-  year = {2016},
-  pages = {1--5}
-}
-
-@article{pollock_automatic_1984,
-  title = {Automatic spelling correction in scientific and scholarly text},
-  volume = {27},
-  issn = {00010782},
-  url = {http://portal.acm.org/citation.cfm?doid=358027.358048},
-  doi = {10.1145/358027.358048},
-  number = {4},
-  urldate = {2020-06-11},
-  journal = {Communications of the ACM},
-  author = {Pollock, Joseph J. and Zamora, Antonio},
-  month = apr,
-  year = {1984},
-  pages = {358--368}
-}
-
-@inproceedings{broder_resemblance_1998,
-  address = {Salerno, Italy},
-  title = {On the resemblance and containment of documents},
-  isbn = {9780818681325},
-  url = {http://ieeexplore.ieee.org/document/666900/},
-  doi = {10.1109/SEQUEN.1997.666900},
-  urldate = {2020-06-11},
-  booktitle = {Proceedings. {Compression} and {Complexity} of {SEQUENCES} 1997 ({Cat}. {No}.{97TB100171})},
-  publisher = {IEEE Comput. Soc},
-  author = {Broder, A.Z.},
-  year = {1998},
-  pages = {21--29}
-}
-
-@inproceedings{dong_efficient_2011,
-  address = {Hyderabad, India},
-  title = {Efficient k-nearest neighbor graph construction for generic similarity measures},
-  isbn = {9781450306324},
-  url = {http://portal.acm.org/citation.cfm?doid=1963405.1963487},
-  doi = {10.1145/1963405.1963487},
-  language = {en},
-  urldate = {2020-06-11},
-  booktitle = {Proceedings of the 20th international conference on {World} wide web - {WWW} '11},
-  publisher = {ACM Press},
-  author = {Dong, Wei and Moses, Charikar and Li, Kai},
-  year = {2011},
-  pages = {577}
-}
-
-@inproceedings{bartolini2002string,
-  title={String matching with metric trees using an approximate distance},
-  author={Bartolini, Ilaria and Ciaccia, Paolo and Patella, Marco},
-  booktitle={International Symposium on String Processing and Information Retrieval},
-  pages={271--283},
-  year={2002},
-  organization={Springer}
-}
-
-@article{damerau1964technique,
-  title={A technique for computer detection and correction of spelling errors},
-  author={Damerau, Fred J},
-  journal={Communications of the ACM},
-  volume={7},
-  number={3},
-  pages={171--176},
-  year={1964},
-  publisher={ACM New York, NY, USA}
-}
-
 @inproceedings{levenshtein1966binary,
   title={Binary codes capable of correcting deletions, insertions, and reversals},
   author={Levenshtein, Vladimir I},
@@ -87,39 +7,6 @@
   number={8},
   pages={707--710},
   year={1966}
-}
-
-@article{dice1945measures,
-  title={Measures of the amount of ecologic association between species},
-  author={Dice, Lee R},
-  journal={Ecology},
-  volume={26},
[…]
