+++ paper/paper.bib
@@ -45,3 +45,65 @@
   year={1980},
   publisher={British Library Research and Development Department London}
 }
+
+@article{postel1969kolner,
+  title={Die K{\"o}lner Phonetik. Ein Verfahren zur Identifizierung von Personennamen auf der Grundlage der Gestaltanalyse},
+  author={Postel, Hans Joachim},
+  journal={IBM-Nachrichten},
+  volume={19},
+  pages={925--931},
+  year={1969}
+}
+
+@book{manning2008introduction,
+  title={Introduction to information retrieval},
+  author={Manning, Christopher D and Sch{\"u}tze, Hinrich and Raghavan, Prabhakar},
+  year={2008},
+  publisher={Cambridge university press}
+}
+
+@book{baeza1999modern,
+  title={Modern information retrieval},
+  author={Baeza-Yates, Ricardo and Ribeiro-Neto, Berthier and others},
+  volume={463},
+  year={1999},
+  publisher={ACM press New York}
+}
+
+@article{fellegi1969theory,
+  title={A theory for record linkage},
+  author={Fellegi, Ivan P and Sunter, Alan B},
+  journal={Journal of the American Statistical Association},
+  volume={64},
+  number={328},
+  pages={1183--1210},
+  year={1969},
+  publisher={Taylor \& Francis}
+}
+
+@book{christen_data_2012,
+  address = {Berlin, Heidelberg},
+  title = {Data {Matching}},
+  isbn = {978-3-642-31163-5 978-3-642-31164-2},
+  url = {http://link.springer.com/10.1007/978-3-642-31164-2},
+  language = {en},
+  urldate = {2017-12-11},
+  publisher = {Springer Berlin Heidelberg},
+  author = {Christen, Peter},
+  year = {2012},
+  doi = {10.1007/978-3-642-31164-2},
+  file = {Peter Christen auth. Data Matching Concepts and Techniques for Record Linkage, Entity Resolution, and Duplicate Detection.pdf:/Users/Yomgui/Zotero/storage/26DNP2AQ/Peter Christen auth. Data Matching Concepts and Techniques for Record Linkage, Entity Resolution, and Duplicate Detection.pdf:application/pdf}
+}
+
+@book{herzog_data_2007,
+  address = {New York ; London},
+  title = {Data quality and record linkage techniques},
+  isbn = {978-0-387-69502-0},
+  language = {en},
+  publisher = {Springer},
+  author = {Herzog, Thomas N. and Scheuren, Fritz and Winkler, William E.},
+  year = {2007},
+  note = {OCLC: ocn137313060},
+  keywords = {Electronic data processing, Quality control},
+  file = {Herzog et al. - 2007 - Data quality and record linkage techniques.pdf:/Users/Yomgui/Zotero/storage/23ZXL9S7/Herzog et al. - 2007 - Data quality and record linkage techniques.pdf:application/pdf}
+}
