### 2020-10-22 4bcda30
Adding mention of benchmark uses Fixes #170
