+++ paper/paper.md
@@ -46,7 +46,7 @@ Those building blocks can then be used to perform and improve the following task
 * Record linkage, entity resolution etc.
 * Natural language processing
 
-Finally, one should note that, being a code library, **Talisman** is able to archive a standardized way to implement some functions and algorithms whose descriptions are known to be somehow unclear, or imprecise, sometimes by consecrating typical implementation used in an industrial context and sometimes by chosing to respect the spirit of the original paper, against a faulty explanation of how the algorithm should behave.
+Finally, this library can also be used to perform some benchmarks of those building blocks, wrt. precision, recall etc. of the fuzzy matches, which is seldom done in the literature because of how hard it can be to find comprehensive archives aggregating many phonetic algorithms, stemmers etc.
 
 # Related works
 
