### 2020-10-22 954319f
Adding BIBLIOGRAPHY.md Related to #172
