+++ BIBLIOGRAPHY.md
@@ -1,0 +1,102 @@
+# Talisman Bibliography
+
+[BibTex file](https://raw.githubusercontent.com/Yomguithereal/talisman/master/paper/algorithms.bib)
+
+> Bartolini, I., Ciaccia, P., & Patella, M. (2002). String matching with metric trees using an approximate distance. International Symposium on String Processing and Information Retrieval, 271–283.
+
+> Beider, A. (2008). Beider-Morse phonetic matching: An alternative to Soundex with fewer false hits. Avotaynu: The International Review of Jewish Genealogy, 24(2), 12.
+
+> Black, P. E. (2004). Ratcliff/Obershelp pattern recognition. Dictionary of Algorithms and Data Structures, 17.
+
+> Bouchard, G., Brard, P., & Lavoie, Y. (1981). FONEM: Un code de transcription phonétique pour la reconstitution automatique des familles saguenayennes. Population (French Edition), 1085–1103.
+
+> Broder, A. Z. (1998). On the resemblance and containment of documents. Proceedings. Compression and Complexity of SEQUENCES 1997 (Cat. No.97TB100171), 21–29. https://doi.org/10.1109/SEQUEN.1997.666900
+
+> Caumanns, J. (1999). A fast and simple stemming algorithm for German words.
+
+> Damerau, F. J. (1964). A technique for computer detection and correction of spelling errors. Communications of the ACM, 7(3), 171–176.
+
+> Dice, L. R. (1945). Measures of the amount of ecologic association between species. Ecology, 26(3), 297–302.
+
+> Dong, W., Moses, C., & Li, K. (2011). Efficient k-nearest neighbor graph construction for generic similarity measures. Proceedings of the 20th International Conference on World Wide Web - WWW ’11, 577. https://doi.org/10.1145/1963405.1963487
+
+> Guth, G. J. (1976). Surname spellings and computerized record linkage. Historical Methods Newsletter, 10(1), 10–19.
+
+> Hamming, R. W. (1950). Error detecting and error correcting codes. The Bell System Technical Journal, 29(2), 147–160.
+
+> Harman, D. (1991). How effective is suffixing? Journal of the American Society for Information Science, 42(1), 7–15.
+
+> Holmes, D., & McCabe, C. M . (2002). Improving precision and recall for soundex retrieval. Proceedings. International Conference on Information Technology: Coding and Computing, 22–26.
+
+> Hood, D. (2002). Caverphone: Phonetic matching algorithm. Technical Paper CTP060902, University of Otago, New Zealand.
+
+> Hood, D. (2004). Caverphone revisited. Technical Paper CTP150804, 10.
+
+> Jaccard, P. (1912). The distribution of the flora in the alpine zone. 1. New Phytologist, 11(2), 37–50.
+
+> Jaro, M. A. (1989). Advances in record-linkage methodology as applied to matching the 1985 census of Tampa, Florida. Journal of the American Statistical Association, 84(406), 414–420.
+
+> Jaro, M. A. (1995). Probabilistic linkage of large public health data files. Statistics in Medicine, 14(5–7), 491–498.
+
+> Jenkins, M.-C., & Smith, D. (2005). Conservative stemming for search and indexing. Proc. ACM SIGIR.
+
+> Khan, S. I., & Hoque, A. S. Md. L. (2016). Similarity analysis of patients’ data: Bangladesh perspective. 2016 International Conference on Medical Engineering, Health Informatics and Technology (MediTec), 1–5. https://doi.org/10.1109/MEDITEC.2016.7835390
+
[…]
