### 2020-10-22 f264fa7
Linking to bibliography in README.md Related to #172
