### 2020-10-22 Yomguithereal/talisman#172 @Yomguithereal
Hello @Fil. I have clarified the paper and added something to express #170. I also added, as per your recommendation, the extensive bibliography of implemented methods to the documentation of the library.
