### 2020-10-22 7a52c85
Adding note on module imports
