### 2020-10-22 #2405 @Yomguithereal
@atanikan I have added a note regarding module imports and `require` [here](https://yomguithereal.github.io/talisman/).

Regarding:

> Then yes, I meant performance benchmarking with other similar libraries. I understand Talisman aims to be an archive/collection of several text-based search/information retrieval techniques. And not exactly a brand new algorithm. I however feel some benchmarking, be it in terms of accuracy or speed of execution by taking a large dataset as an example, may prove to be useful for wider use and a greater acceptance of your library.

There exists some speed benchmarks regarding some functions of the library (notably Levenshtein distance, because this library was used as an experimental starting point for the creation of many very fast implementation of the distance in JavaScript) but for most of the method there is nothing to benchmark against because nobody implemented those functions.

Regarding precision/recall etc. the issue here is that the relevant databases that could enable such benchmarks don't exist yet because most of those methods were created as heuristics geared toward the resolution of particular problems and so it made no sense to most of their authors to compare them to others. This said, I believe there is a point to be made regarding some classes of functions present in the library such as phonetic algorithms geared towards proper names and their precision/recall. I personally intend to develop such a benchmark and I have started created a database that could be used for this purpose (along with another phonetic algorithm I am developing right now), but unfortunately this is an ongoing work which won't be finished anywhere soon.

This said, do you want me to, at least, try to compile some speed benchmarks regarding a subset of the library's functions that have some other JavaScript implementations?
