### 2020-10-22 #2405 @atanikan
> This said, do you want me to, at least, try to compile some speed benchmarks regarding a subset of the library's functions that have some other JavaScript implementations?

It's fine. I see the article does speak to the fact that one can benchmark if they would like. Since the article nor the library makes an outright claim to have the "fastest"/"best". I suppose we can benchmarking stats for now. Appreciate the changes made in the documentation as well. @kakiac My review is complete as well. Thanks, @Yomguithereal. Fantastic job.
