### 2020-11-05 #2405 @kakiac
**Editor's "After reviewers recommend acceptance" checks:**
 - [ ] Get a new proof with the @whedon generate pdf command.
 - [ ] Download the proof, check all references have DOIs, follow the links and check the references.
 - [ ] Whedon can help check references with the command @whedon check references
 - [ ] Proof-read the paper and ask authors to fix any remaining typos, badly formed citations, awkward wording, etc..
 - [ ] Ask the author to make a tagged release and archive, and report the version number and archive DOI in the review thread.
 - [ ] Check the archive deposit has the correct metadata (title and author list), and request the author edit it if it doesn’t match the paper.
 - [ ] Run @whedon set <doi> as archive.
 - [ ] Run @whedon set <v1.x.x> as version if the version was updated.
 - [ ] Run @whedon accept to generate the final proofs, which has Whedon notify the @openjournals/joss-eics team that the paper is ready for final processing.
 - [ ] Mention EiC to take over
