### 2020-11-05 0cfa010
Adding missing DOIs
