+++ paper/paper.bib
@@ -17,7 +17,8 @@
   number={2},
   pages={37--50},
   year={1912},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/j.1469-8137.1912.tb05611.x}
 }
 
 @article{odell1956profit,
@@ -104,6 +105,7 @@
   author = {Herzog, Thomas N. and Scheuren, Fritz and Winkler, William E.},
   year = {2007},
   note = {OCLC: ocn137313060},
+  doi = {10.1007/0-387-69505-2},
   keywords = {Electronic data processing, Quality control},
   file = {Herzog et al. - 2007 - Data quality and record linkage techniques.pdf:/Users/Yomgui/Zotero/storage/23ZXL9S7/Herzog et al. - 2007 - Data quality and record linkage techniques.pdf:application/pdf}
 }
