+++ paper/algorithms.bib
@@ -130,7 +130,8 @@
   number={2},
   pages={37--50},
   year={1912},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/j.1469-8137.1912.tb05611.x}
 }
 
 @article{winkler1990string,
