### 2020-11-05 #2405 @Yomguithereal
I pushed a commit adding the missing DOIs to the paper's bibliography.
