+++ paper/paper.bib
@@ -79,7 +79,10 @@
   number={328},
   pages={1183--1210},
   year={1969},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi = {10.1080/01621459.1969.10501049}, 
+  URL = {https://www.tandfonline.com/doi/abs/10.1080/01621459.1969.10501049},
+  eprint = {https://www.tandfonline.com/doi/pdf/10.1080/01621459.1969.10501049}
 }
 
 @book{christen_data_2012,
