+++ BIBLIOGRAPHY.md
@@ -32,7 +32,7 @@
 
 > Hood, D. (2004). Caverphone revisited. Technical Paper CTP150804, 10.
 
-> Jaccard, P. (1912). The distribution of the flora in the alpine zone. 1. New Phytologist, 11(2), 37–50.
+> Jaccard, P. (1912). The distribution of the flora in the alpine zone. 1. New Phytologist, 11(2), 37–50. https://doi.org/10.1111/j.1469-8137.1912.tb05611.x
 
 > Jaro, M. A. (1989). Advances in record-linkage methodology as applied to matching the 1985 census of Tampa, Florida. Journal of the American Statistical Association, 84(406), 414–420.
 
