### 2020-11-05 #2405 @kakiac
@Yomguithereal apologies - I should have been clearer - Upon successful completion of the review, you need to make a tagged release of the software (via github, see here for info https://git-scm.com/book/en/v2/Git-Basics-Tagging), and deposit a copy of the repository with a data-archiving service such as Zenodo or figshare.

You then need to get the DOI you will be assigned from them (it is automatically assigned once you create a repository).

As soon as you have both copy the release version number (from github) and the DOI in this thread.

See here for more info under section **The review process** https://joss.readthedocs.io/en/latest/submitting.html#submitting-your-paper - let me know if you need any help!
