### 2020-11-05 #2405 @Yomguithereal
@kakiac the release is `1.1.3-zenodo` and the attached DOI is `10.5281/zenodo.4247869` (but the [https://doi.org/10.5281/zenodo.4247869](https://doi.org/10.5281/zenodo.4247869) url does not yet work as I write this message. I guess it is just Zenodo lag?).
