### 2020-11-05 #2405 @kakiac
Yes, no worries it worked for me just now, I think it takes some time to get registered in the doi database, I think. Thanks for doing that!
