### 2020-11-05 #2405 @kakiac
> @kakiac Please wait to ping openjournals/joss-eics until a submission is ready to be taken over, if possible. Just saves a little bit of email traffic for us. Your last checkbox isn't needed since that happens with `@whedon accept`.

Oh apologies! I thought I needed to get that tagged if accepted by reviewers! Will bear that in mind in the future!
