### 2020-11-09 #2405 @kakiac
@Yomguithereal not sure if you spotted my message above - it seems that the authors list of your JOSS paper and the zenodo archive does not match. Can you please make sure they align?
