### 2020-11-09 #2405 @kthyng
Hi @Yomguithereal! We also prefer that the title in your Zenodo archive matches your JOSS paper exactly. Would you want to update that?
