### 2020-11-09 #2405 @kthyng
Paper looks good! We can proceed after hearing about the Zenodo archive.
